Add basement catch-all classes
The basement::rootrecipe class is meant to be THE entry point of
projects that are based on the basement recipes. It does all the
required setup and makes the inheriting recipe a root recipe.

The basement::buildall class is just a helper to build everything that
is provided by the basement recipes. It is not intended to be used by
other projects but as a helper for the CI system.