Relicense as MIT
The sole use of the GPL-3.0-or-later license prohibited the use of the
Basement project in other projects and left a gray area when patches are
applied to other software that is licensed under an incompatible
license.

Included patches are covered by the licence of the affected software.
This exception is taken from Buildroot where it did stand the test of
time.

Closes #7.

Signed-off-by: Jan Klötzke <jan@kloetzke.net>
Signed-off-by: Philipp Trommler <github@philipp-trommler.me>