host-compat-toolchain: do not set LD_LIBRARY_PATH when used
The host compat toolchain comes with a number of shared libraries that
are needed when being executed. Unfortunately the same set of libraries
may be present on the host/sandbox. Now if an executable from the
host/sandbox is started it will load the wrong library (observed with
msgmerge which tries to load libgomp.so.1).

We instead wrap all executables with a tiny shell script that sets
LD_LIBRARY_PATH just for them. Another option would have been to use
the $ORIGIN mechanism of glibc in the RPATH. But the toolchain has
several executables outside of usr/bin which all would need a different
RPATH.