devel/cmake: add version 3.17.3
CMake is quite interesting in that can be built in two ways. It can
bootstrap itself where all dependencies are included in the source
tree. This is used to build an initial cmake.

The real version uses cmake to build itself and which has a number of
dependent libraries which in again need cmake (jsoncpp -> meson -> ninja
-> cmake). All these packages need to be built with the boostrap cmake.