Rewire toolchain variable handling
Attach almost all toolchain relevant variables to the toolchain itself.
This has the benefit that the variables are updated depending on the
usage of a particular toolchain. This is important for the upcoming
tool remapping to have a consistent tool+variable setting on the
consuming recipes.

The host defaults are still set in default.yaml so that a consistent
host build is possible.