cpackage: handle static host builds
On host builds everything is built statically. Unfortunately some
packages (python3-minimal in particular) will still build shared
libraries. Depending on the actual code and linking this might lead to
linker error if the relocations cannot be resolved. To fix that we
always build with "-fPIC". This will produce marginally slower code but
the host binaries are not the end result so that should not matter.