devel/libtool: split into the generic scripts and the c-library
The generic libtool scripts are built manually to keep them untainted
from the target toolchain. These are just shell scripts and m4-files so
they don't depend on the target architecture. Unfortunately the
included build system does not allow to disable the c-library. That's
why we roll our own "build" for the libtoolize script.

At the same time we make the script relocatable. There were only a few
absolute paths in the script anyway.