install: better host compilation support
When building for the host we want to use only static libraries. Most
packages obey the respective switches but there are still a couple that
either do not have static/shared switches are build both
unconditionally. For such packages the install class makes sure to
install only static library on host builds.

We still link executables without "-static", though. The reason is glibc
which cannot reasonably be linked statically. The install class will
therefore need to fix pkg-config files. The private dependencies that
are only used when linking statically must be promoted to regular
dependencies. For cross builds nothing is changed.