classes: introduce BASEMENT_LIBS switch
The BASEMENT_LIBS switch can be set to control the build type and
installation of libraries. If unset, the default is dependent on the
overall built type as before: build shared libs on cross-compilation
and only static libs for host builds.

The switch might be set to one of the following values to override this
default:

 * shared - build and install shared libraries. Static libraries will
            still be installed if present, though.
 * static - build and install static libraries only. Shared libraries
            will never be installed.
 * both - Build and install both shared and static libraries.