devel/compat/gcc: apply libtool patch
Because gcc does not use the autotools class it did not automatically
get the libtool patch. This leads to cross compile linking errors when
cross-compiling a native compiler (the last step of
devel::host-compat-toolchain). OTOH we cannot use the libtool class
directly because there are patched ltmain.sh files in the sources.
Instead we apply the matching patch directly.