devel/gcc: apply libtool patch
The gcc recipe does not use the autotools class that pulls in the
libtool patch. We have to do that manually. Otherwise there are some
subtle linking problems in the final stage of the gcc-native built that
are not immediately visible because some host libraries are used where
they shouldn't.