devel/make: do not rely on host make
Build our make version as host tool and use it everywhere. There are
subtle problems regarding the interoperability of make with the Bob
internal job server. Only recent versions of make use non-blocking
pipes like Bob. Because we cannot know what is on the host we make sure
that we always use our version.