devel/pkg-config: use self-contained build as tool
The build of pkg-config with separate glib requires pkg-config itself.
This was not a problem yet but with glib moving to meson it will cause
problems. When used as tool we rely on the built-in version of glib to
satisfy the dependency. When built as regular package glib will be used
as normal dependency still.