libs/ncurses: fix downstream build failures
The wrong kind of libraries were built by default. Instead of shared
libraries static ones were built and vice versa. This lead to link
errors downstream because of wrong relocations.

Some packages will also still need the original library name without the
"w" suffix. The solution commonly used is to add a symlink to provide
the old name too. On the target file system only the "w" variant shared
library is installed because that's the soname in any case.

Fixes #58.