python/python3: always extract sources from scratch
The patches that are applied on the sources are stacked on each other
and fail the "reverse apply" check of the patch class. This leads to
build failures whenever the checkout step is re-run. We now start over
from scratch every time in the hope that no user will ever accidentally
edit the python sources...