libs/openssl: add fixup for static builds
Static library builds of libcrypto need to explicitly link with libz.
The produced pkgconfig file is just using the '-lz' option which assumes
that the library is in the linker search path. This is not true for our
build system. Fix it by adding a dependency to the libz pc file instead.
This will make pkgconfig emit the needed additional library path.