core/util-linux: Expose tools
This exposes "/usr/sbin" to the set of tools that can be used in other
recipes. Although "/usr/bin" includes binaries, as well, those are
mainly interesting for the actual target. In "/usr/sbin", on the other
hand, is e.g. mkfs that is of particular interest when building images.