Rework {C,CPP,CXX,LD}FLAGS handling
Sometimes the user wants to add specific options to the various compiler
flags variables. Therefore it's not a good idea to hard-code the flags
in the toolchain. Instead, provide them as normal variables so that the
usual environment handling works for these variables.

There is a special exception for the "host-compat-toolchain", though.
This toolchain is intended to be assigned as "target-toolchain" to build
selected packages for the host when a different target toolchain is
currently active. To prevent any interference of the currently active
flags, they are forced by "host-compat-toolchain". Otherwise changing
the flags could unintentionally affect host tool builds.

Fixes #178.