devel::git - Install using symlinks
git by default uses hardlinks to provide multiple executables with the
same binary. This however doesn't translate well when copying around the
installed files during the packaging step as the hardlinks are not
reestablished, but the files are now duplicated, which results in a git
installation with a size of more than 300MB.

For these cases git build system has a make variable called
NO_INSTALL_HARDLINKS which can be passed along during make install to
use symlinks instead of hardlinks.