devel/cross-toolchain: fix build
The devel::cross-toolchain and devel::gcc recipes incorrectly used the
host-toolchain to build the compiler. This is not correct if some other
cross compiling toolchain is active, like the
devel::host-compat-toolchain. In this case we must do a proper canadian
cross toolchain build.

To optimize the candian cross build, only the compiler itself is built
as canadian cross. The rest is reused from the first stage cross
compiler because it should be identical.

Fixes #189.