autoconf: work around perl file timestamp issues
Perl apparently has poor support for sub-second file timestamps. In
particular, the mtime is truncated if a file is moved across file
systems. This can happen with different autotools (autoheader in
particular) because they usually create their files in a temporary
directory and then move the result to the final location.

If /tmp is on a different file systems, the time stamp will be
truncated. This can lead to the situation where "config.h.in" appears to
be older than aclocal.m4 because the timestamp of the former is
truncated. This in turn can trigger another round of autoconf invocation
at build time, which will fail in a sandbox.

Work around this by using the workspace as temporary directory. Then the
rename() system call will work, naturally retaining the full timestamp
resolution.