patch: improve tracking of applied patches
Each applied patch is tracked individually. This has the benefit of
being able to handle situations where multiple patches touch the same
file.

Also remove the check if the patch is still applied. This proved to be
unreliable, especially if multiple patches touch the same file. We now
solely rely on the tracked applied patches and compare them to the
desired ones. If they match, the class will assume the patch is still
applied.

Fixes #55. Fixes #180.