install: better handling of shared libs target install
When installing shared libraries for the target file system, only the
shared object itself should be copied. If the target system uses library
versioning (like most do), to SONAME of the shared library might differ
from the actual file name. Create a symlink in this case, so that the
dynamic loader finds the library.

This approach is much more reliable. It will work for systems or
libraries without versioning and also does not rely on certain naming
schemes.