basement/update-config: auto update GNU config helpers
Each autotools package brings its own copy of config.guess and
config.sub. Some of them have not been updated in years. To use the
latest version of these files consistently, add a helper class that
automatically updates them. The autotools class pulls it in
for obvious reasons.

It is realized as separate class because there are some packages that
use the GNU config helpers without the rest of the autotools. A prime
example is perl which naturally does not want to pull in the rest of the
autotools class.