add llvm v19.1.4
LLVM includes several libs as well as some tools. Compiling
those libs without the llvm-tools seams to be not supported.

The llvm-recipe therefor defines tools and lib-multipackages.

Also a clang-class is added which connects the clang-compiler to the
target-toolchains sysroot. This class can be inherited by packages where
clang should be used. The class also offers to disable clang treewide by
not setting BASEMENT_HAS_CLANG to true.