gcc: add option to disable libada build
Libada is the runtime library needed to run ada applications on different
systems. It is enabled by default if `ada` is an enabled language. But
o build libada a crt0 is needed which is not available if a newlib
based toolchain is build, so linking libada failes. Add a switch to
disable it in these cases. A separate (bare) runtime must be used for
bare metal ada.