config: fix potential crash
Depending on the loading order, some recipes may not have a "Config"
property. This happens if a layer is parsed before the basement layer.
Naturally, these recipes cannot utilize the "Config" keyword but the
plugin must still cope with that because the "onFinish" method is
potentially invoked on such recipes.