python3: restructure classes
Put all classes that handle the different types of Python build systems
into a common namespace. At the same time, split the setuptools and pip
methods from one another.

The python3::cext module is converted to a mixin class. Whether or not a
Python module has a C-extension is independent from the underlying build
system.