install: add function to fix shebang line
Using an absolute path for the interpreter breaks if the interpreter is
installed somewhere else. Add a helper function to replace the absolute
path by '/usr/bin/env <interpreter>'.

In case the interpreter is called with arguments, `env -S` must be used.
This only works for coreutils-env and is not supported by busybox 'env'.