cross-toolchain: allow setting -march and -mcpu
The toolchain provides the default CFLAGS and CXXFLAGS. On the one hand,
the compiler has a default architecture and CPU model. But sometimes the
project may want to select a different architecture or CPU. This is done
by the -march= and -mcpu= flags. Providing some hooks with
CROSS_TOOLCHAIN_ARCH and CROSS_TOOLCHAIN_CPU to set these flags when the
toolchain is brought in:

  depends:
      - name: devel::cross-toolchain-x86_64-linux-gnu
        use: [tools, environment]
        forward: True
        environment:
            CROSS_TOOLCHAIN_ARCH: skylake

This will add "-march=skylake" to the CFLAGS and CXXFLAGS. More
importantly, this will enable us later to extract the selected features
(e.g. AVX) of the target architecture/CPUs.