cross-toolchain: provide feature flags for Arm, AArch64 and x86
Some packages need to know which features the targeted system/CPU has.
This depends on the compiler settings in the first place. We now have
all the information about the selected model. Based on the selected gcc
and CFLAGS configuration, the available features can be extracted.

They are provided as different CPU_HAS_* toolchain variables. There are
different variables for the different architectures. Define the most
commonly used for the moment. This schema can be extended in the future
as needed.