Rectify default C(++) toolchain variables
Setting the default C(++) toolchain tool names in default.yaml is
dangerous. They are then set all the time even if not target toolchain
is used. This is especially true for AUTOCONF_HOST which should better
not be set at all. Otherwise we could accidentally indicate cross
compilation if the target-toolchain is not used but the (old) definition
is retained from default.yaml.