host-compat-toolchain: align LDFLAGS with default
The deve::host-compat-toolchain package defines a regular
"target-toolchain" and the special "host-compat-toolchain" tool. The
former is configured by default in the basement::rootrecipe class and is
used for a number of host tools. Being a target-toolchain, the
C/CXX/LDFLAGS are taken from the environment which is configured in
default.yaml in this case.

On the other hand, the host-compat-toolchain is used to replace the
current toolchain wherever needed. It therefore overwrites any existing
(C,CXX,LD)FLAGS. We have to make sure that the overwritten LDFLAGS have
the same value as the default from default.yaml. Otherwise we create
different variants if the same recipe is compiled with the default
target-toolchain of deve::host-compat-toolchain.