devel::cross-toolchain: add clang support
The basement project is relying on gcc for the main compiler. Adding
clang to the mix does not change its fundamentally. That is, even with
clang the binutils and major low level libraries (libgcc, libstdc++) are
still used.

To make things easy and to have only one package, put the clang wrapper
scripts inside the gcc cross toolchain. The rest is configuration.
The -clang multiPackage provides the same toolchains but replaces gcc
and c++ with clang. Using these variants assumes that the "clang" tool
is available and it will be added automatically to the dependencies for
every package that is using the cross target-toolchain.