autoconf: add option to run aclocal
Apparently, autoreconf does not update any aclocal macros even if they
are available (devel::autoconf-archive). If a recipe needs updates of
these macros, it has to pass the -a option to autoconfReconfigure()
which runs "aclocal -I m4 --install" before autoreconf. Usually this is
not needed so it has to be a selective opt-in.

Optionally, the arguments to aclocal can be overridden by "-A".