libs::openssl: split into lib and binary package
Right now the openssl binary is always installed on the target. If
openssl is an indirect dependency of another package this isn't always
desired. Therefore, split up the binary package from the lib package.

If one wants to have the openssl binary, libs::openssl should be added
to the dependencies.

NOTE: This means the openssl binary is removed from the sandbox. If a
package needs the openssl binary as a host tool, the new added "openssl"
tool have to be added to the buildTools.