install: fix shebang mangling
The implementation of the shebang mangling had several flaws. First of
all, it did not do anything because the "find" invocation was broken.

Also, the kernel will strip any number of spaces and tabs after the
"#!" marker. We need to do the same when comparing the command with
known good values. Additionally, that list was missing the /bin/sh and
/bin/bash commands that we certainly do not want to replace.