python::python3: fix dev package
The python-dev package holds a magic sysconf module that is used by
setuptools and friends to find crucial paths. In particular, the path
to the include files is taken from there.

So far, the module did hold absolute paths that pointed to the intended
install location. That is completely wrong because the package never
ends up there and needs to be relocatable. Patch the module accordingly
to rewrite all absolute paths to one relative to the sysconfig module.