devel::llvm: fix cross compilation
When cross-compiling llvm, the devel::llvm-host-tools package is built
with the host-compat-toolchain. In contrast to other packages, the llvm
tools are still linked shared. That is, libLLVM.so and friends need to
be available when executing these tools. Fortunately, the llvm build
system sets a proper RPATH so that we just have to install the shared
libraries.